// src/components/flex/Flex.type.ts
export enum FlexDirection {
  Row = 'row',
  Column = 'column',
}

export enum FlexJustify {
  Start = 'flex-start',
  End = 'flex-end',
  Center = 'center',
  Between = 'space-between',
  Around = 'space-around',
}

export enum FlexAlign {
  Start = 'flex-start',
  End = 'flex-end',
  Center = 'center',
  Stretch = 'stretch',
  Baseline = 'baseline',
}

export enum FlexWrap {
  Wrap = 'wrap',
  Nowrap = 'nowrap',
}